recommend/requirements.txt
coverage==3.5.1
ipython==0.12
nose==1.1.2
numpy==1.6.1
scikits.crab==0.1
scikits.learn==0.8.1
scipy==0.10.0
